requirements.txt
cycler==0.10.0
Cython==0.29.11
decorator==4.4.0
future==0.17.1
grakel-dev==0.1a5
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.3
nose==1.3.7
numpy==1.16.4
pandas==0.24.2
pyparsing==2.4.0
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn-pandas==1.8.0
